Compute averages as floating-point.

+++ jitsi-utils/src/main/java/org/jitsi/utils/queue/QueueStatistics.java
@@ -117,8 +117,8 @@ public class QueueStatistics implements PacketQueue.Observer
         double duration = (now - firstPacketAddedMs) / 1000d;
         stats.put("duration_s", duration);
         stats.put("average_remove_rate_pps", totalPacketsRemoved.sum() / duration);
-        stats.put("average_queue_size_at_remove", totalSize.sum() / totalPacketsRemoved.sum());
-        stats.put("average_queue_wait_time", totalWait.sum() / totalPacketsRemoved.sum());
+        stats.put("average_queue_size_at_remove", totalSize.sum() / (double)totalPacketsRemoved.sum());
+        stats.put("average_queue_wait_time", totalWait.sum() / (double)totalPacketsRemoved.sum());
 
         return stats;
     }
